Something wrongs in cq code. Poke cq code was create previous using reply cq code, now is poke cq code. Reply cq code should be use 'long' instead of 'int'.

+++ src/main/java/com/github/cao/awa/apricot/message/element/cq/factor/replay/CqReplyFactor.java
@@ -11,7 +11,7 @@ public class CqReplyFactor extends CqFactor {
     @Override
     public MessageElement create(ApricotServer server, Map<String, String> args) {
         return new ReplyMessageElement(
-                Integer.parseInt(args.get("id"))
+                Long.parseLong(args.get("id"))
         );
     }
